+++ enterprise/internal/completions/httpapi/handler.go
@@ -29,15 +29,14 @@ func newCompletionsHandler(rl RateLimiter, traceFamily string, getModel func(typ
 		ctx, cancel := context.WithTimeout(r.Context(), maxRequestDuration)
 		defer cancel()
 
-		completionsConfig := client.GetCompletionsConfig(conf.Get().SiteConfig())
-		if completionsConfig == nil || !completionsConfig.Enabled {
-			http.Error(w, "completions are not configured or disabled", http.StatusInternalServerError)
-			return
-		}
-
 		if isEnabled := cody.IsCodyEnabled(ctx); !isEnabled {
 			http.Error(w, "cody experimental feature flag is not enabled for current user", http.StatusUnauthorized)
 			return
+		}
+
+		completionsConfig := client.GetCompletionsConfig(conf.Get().SiteConfig())
+		if completionsConfig == nil {
+			http.Error(w, "completions are not configured or disabled", http.StatusInternalServerError)
 		}
 
 		var requestParams types.CompletionRequestParameters